Use subscription id from config
